+++ pkg/provider/prediction_environment_resource.go
@@ -6,12 +6,26 @@ import (
 	"fmt"
 
 	"github.com/datarobot-community/terraform-provider-datarobot/internal/client"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+const (
+	batchJobPriorityLow    int = 0
+	batchJobPriorityMedium int = 1
+	batchJobPriorityHigh   int = 2
+
+	managedBySelfManaged     string = "selfManaged"
+	managedByManagementAgent string = "managementAgent"
+	managedByDatarobot       string = "datarobot"
 )
 
 // Ensure provider defined types fully satisfy framework interfaces.
@@ -59,6 +73,48 @@ func (r *PredictionEnvironmentResource) Schema(ctx context.Context, req resource
 					stringplanmodifier.RequiresReplace(),
 				},
 			},
+			"batch_jobs_max_concurrent": schema.Int64Attribute{
+				Optional:            true,
+				MarkdownDescription: "The maximum number of concurrent batch prediction jobs.",
+			},
+			"batch_jobs_priority": schema.StringAttribute{
+				Optional: true,
+				Validators: []validator.String{
+					stringvalidator.OneOf("low", "medium", "high"),
+				},
+				MarkdownDescription: "The importance of batch jobs.",
+			},
+			"supported_model_formats": schema.ListAttribute{
+				Optional:            true,
+				ElementType:         types.StringType,
+				MarkdownDescription: "The list of supported model formats.",
+				PlanModifiers: []planmodifier.List{
+					listplanmodifier.RequiresReplace(),
+				},
+			},
+			"managed_by": schema.StringAttribute{
+				Optional:            true,
+				Computed:            true,
+				Default:             stringdefault.StaticString(managedBySelfManaged),
+				MarkdownDescription: "Determines if the prediction environment should be managed by the management agent, datarobot, or self-managed. Self-managed by default.",
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+			},
+			"credential_id": schema.StringAttribute{
+				Optional:            true,
+				MarkdownDescription: "The ID of the credential associated with the data connection. Only applicable for external prediction environments managed by DataRobot.",
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+			},
+			"datastore_id": schema.StringAttribute{
+				Optional:            true,
+				MarkdownDescription: "The ID of the data store connection configuration. Only applicable for external prediction environments managed by DataRobot.",
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+			},
 		},
 	}
 }
@@ -87,11 +143,7 @@ func (r *PredictionEnvironmentResource) Create(ctx context.Context, req resource
 	}
 
 	traceAPICall("CreatePredictionEnvironment")
-	createResp, err := r.provider.service.CreatePredictionEnvironment(ctx, &client.CreatePredictionEnvironmentRequest{
-		Name:        data.Name.ValueString(),
-		Description: data.Description.ValueString(),
-		Platform:    data.Platform.ValueString(),
-	})
+	createResp, err := r.provider.service.CreatePredictionEnvironment(ctx, buildPredictionEnvironmentRequest(data))
 	if err != nil {
 		resp.Diagnostics.AddError("Error creating Prediction Environment", err.Error())
 		return
@@ -134,6 +186,7 @@ func (r *PredictionEnvironmentResource) Read(ctx context.Context, req resource.R
 	if predictionEnvironment.Description != "" {
 		data.Description = types.StringValue(predictionEnvironment.Description)
 	}
+	data.ManagedBy = types.StringValue(predictionEnvironment.ManagedBy)
 
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
 }
@@ -147,12 +200,7 @@ func (r *PredictionEnvironmentResource) Update(ctx context.Context, req resource
 	}
 
 	traceAPICall("UpdatePredictionEnvironment")
-	_, err := r.provider.service.UpdatePredictionEnvironment(ctx,
-		data.ID.ValueString(),
-		&client.UpdatePredictionEnvironmentRequest{
-			Name:        data.Name.ValueString(),
-			Description: data.Description.ValueString(),
-		})
+	_, err := r.provider.service.UpdatePredictionEnvironment(ctx, data.ID.ValueString(), buildPredictionEnvironmentRequest(data))
 	if err != nil {
 		if errors.Is(err, &client.NotFoundError{}) {
 			resp.Diagnostics.AddWarning(
@@ -188,4 +236,54 @@ func (r *PredictionEnvironmentResource) Delete(ctx context.Context, req resource
 
 func (r *PredictionEnvironmentResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
+}
+
+func buildPredictionEnvironmentRequest(data PredictionEnvironmentResourceModel) *client.PredictionEnvironmentRequest {
+	request := &client.PredictionEnvironmentRequest{
+		Name:                       data.Name.ValueString(),
+		Description:                data.Description.ValueString(),
+		Platform:                   data.Platform.ValueString(),
+		ManagedBy:                  data.ManagedBy.ValueString(),
+		IsManagedByManagementAgent: isManagedByManagementAgent(data.ManagedBy.ValueString()),
+	}
+
+	if len(data.SupportedModelFormats) > 0 {
+		supportedModelFormats := make([]string, len(data.SupportedModelFormats))
+		for i, v := range data.SupportedModelFormats {
+			supportedModelFormats[i] = v.ValueString()
+		}
+		request.SupportedModelFormats = supportedModelFormats
+	}
+	if IsKnown(data.BatchJobsPriority) {
+		request.Priority = priorityStringToEnum(data.BatchJobsPriority.ValueString())
+	}
+	if IsKnown(data.BatchJobsMaxConcurrent) {
+		batchJobsMaxConcurrent := data.BatchJobsMaxConcurrent.ValueInt64()
+		request.MaxConcurrentBatchPredictionsJob = &batchJobsMaxConcurrent
+	}
+	if IsKnown(data.CredentialID) {
+		request.CredentialID = data.CredentialID.ValueStringPointer()
+	}
+	if IsKnown(data.DatastoreID) {
+		request.DatastoreID = data.DatastoreID.ValueStringPointer()
+	}
+
+	return request
+}
+
+func priorityStringToEnum(priority string) *int {
+	var priorityValue int
+	switch priority {
+	case "high":
+		priorityValue = batchJobPriorityHigh
+	case "medium":
+		priorityValue = batchJobPriorityMedium
+	default:
+		priorityValue = batchJobPriorityLow
+	}
+	return &priorityValue
+}
+
+func isManagedByManagementAgent(managedBy string) bool {
+	return managedBy == managedByManagementAgent || managedBy == managedByDatarobot
 }